Create MUST READ!
created a MUST READ file to address the issue of two files being larger than 100mb, which hinders pushing to github. 

free git LFS says 1GiB free, but since there is five of us and later the TA's, the limit will be reached fast. If no changes are made to it, like i suspect, then git LFS will be fine i think. 

NEED TO DISCUSS THE MATTER WITH TAs and LECTURER AS IT STATES ONLY THAT WHICH IS IN GITHUB WILL BE ASSESSED!