Add Attribution.txt and edit gitignore
A rather brunt way of editing gitignore. I took out the files from the project, because i had already commited, and then added them to gitignore. Once i commit this, then when i add them back, it should get ignored (apparently if they stay in the repo while i edit gitignore it wont ignore).

Added an attribution file for all the assets we are going to use in the game.