Updating gitignore file in main
This is just in case someone has put the files in their project already and try to branch from main, they have to keep updating the gitignore file in each new branch.

I will take the liberty of updating the main branch's gitignore so that each new branch can ignore the files.