Some small edits ahead of starting Sprint 2
there were duplicate tags of coin, levelender, life, playerui (didnt really affect the game but resolved it to be safe).
Made some edits to UI Manager and StageScore manager scripts because although ti worked fine in editor, when trying it out on a build, the subtle timings caused errors in finding objects with tags, esp the player. -> made it a direct link to be done in editor.