Some edits made after todays lab
The reason pvp didnt work was override was missing, i had deleted it due to a message in unity editor when i was putting comments on the code but it wasnt because of override but actually because the parameter was overlapping with unity Player built in. Adjusted it and should work now.
Also changed so that boundaries at edge of screen. Change MC swordhitbox prefab to 0.1 duration instead of 0.5 to pseudo solve the trailing problem. To be discussed more